requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
python-dotenv
pydantic
langchain-google-genai
